lib/dnssdint: disable rotation support
When rotation support is enabled, the client assumes the server knows how to
rotate video, which is not currently the case in this project for all
streams.  By disabling this bit, the client performs rotation before sending
video to the server.  Note that this does not keep devices from rotating
their screens, just where the rotation happens.

Signed-off-by: Derrick Lyndon Pallas <derrick@pallas.us>