renderers: reset timing on excessive delay
In my experience, the delay number is usually a small negative value.  When
it gets too large in the positive direction, the system will hang waiting to
render the buffer.

It is not clear to me why the delay is becoming excessive but there are a
couple of reliable ways I've found to reproduce it.  With an iPhone 7+, I

	1. mirrored the phone,
	2. started some music,
	3. put the phone to sleep,
	4. after 15 seconds, woke the phone up.

Prior to this patch, the mirror would hang.  After this patch, it doesn't.

Note that resetting the clock on any delay is not helpful, as it leads to
jitter in streams that should legitimately be delayed.  In the bad case I
was seeing delays in excess of 1M and in the working case I was seeing
legitimate delays of around 50k, so it seemed reasonable to make the reset
threshold 100k.  This could probably be tuned better.

Signed-off-by: Derrick Lyndon Pallas <derrick@pallas.us>